+++ src/components/feature/module/ControlPanelComponent.tsx
